# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InoculAI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/intake": {
      "post": {
        "summary": "Intake",
        "operationId": "intake_intake_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analyze": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/engage/start": {
      "post": {
        "summary": "Engage Start",
        "operationId": "engage_start_engage_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/engage/step": {
      "post": {
        "summary": "Engage Step",
        "operationId": "engage_step_engage_step_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/train/start": {
      "post": {
        "summary": "Train Start",
        "operationId": "train_start_train_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/train/message": {
      "post": {
        "summary": "Train Message",
        "operationId": "train_message_train_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/train/end": {
      "post": {
        "summary": "Train End",
        "operationId": "train_end_train_end_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EngagementStart": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_id"
        ],
        "title": "EngagementStart"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntakeInput": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "synthetic"
        ],
        "title": "IntakeInput"
      },
      "Message": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "email",
              "chat"
            ],
            "title": "Channel"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sender": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Sender"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "channel",
          "text",
          "sender",
          "timestamp"
        ],
        "title": "Message"
      },
...
      "SessionInput": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "SessionInput"
      },
      "TextInput": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "synthetic"
        ],
        "title": "TextInput"
      },
      "TrainingInput": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "synthetic",
          "session_id"
        ],
        "title": "TrainingInput"
      },
      "TrainingStart": {
        "properties": {
          "dna_id": {
            "type": "string",
            "title": "Dna Id"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dna_id",
          "consent"
        ],
        "title": "TrainingStart"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}